CODE/requirements.txt
flask
torch
pandas
transformers
scikit-learn
gensim
faiss-cpu
pyarrow
bert_score
